Project - dependency injection with inheritance Java
This was a project to show dependeny injection. It aim of the project
was to add people to an event, obtain a message of "welcome to the
event" when they register. At the end of the event, all participants
obtained a message of "Thanks for attending".
Each person atteding the event were employees of a company with various
attributes (used inheritance).